+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/ConnectionSupplier.java
@@ -49,13 +49,6 @@ public class ConnectionSupplier implements SqlConnectionSupplier {
         return delegate.get();
     }
 
-    /**
-     * Responsibility of the consumer to close this connection when finished.
-     */
-    public SqlConnection getNewUnsharedConnection() {
-        return delegate.get();
-    }
-
     @Override
     public synchronized void close() {
         if (sharedConnection != null) {